# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Imparables SA - Sistema de gestión naturista
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteResponse {
  message: string;
}

export interface Vendedor {
  idVendedor: number;
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
}

export interface CreateVendedorBody {
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
}

export interface Direccion {
  idDireccion: number;
  calle: string;
  numero: string;
  colonia: string;
  ciudad: string;
  /** @nullable */
  latitud?: number | null;
  /** @nullable */
  longitud?: number | null;
}

export interface ClienteConDireccion {
  idCliente: number;
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
  registrado: string;
  direccion?: Direccion;
}

export interface CreateClienteBody {
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
  registrado: string;
  calle: string;
  numero: string;
  colonia: string;
  ciudad: string;
  /** @nullable */
  latitud?: number | null;
  /** @nullable */
  longitud?: number | null;
}

export interface Producto {
  idProducto: number;
  nombreProducto: string;
  precio: number;
}

export interface CreateProductoBody {
  nombreProducto: string;
  precio: number;
}

export interface DetalleVenta {
  idDetalle: number;
  idProducto: number;
  nombreProducto: string;
  cantidad: number;
  subtotal: number;
}

export interface VentaConDetalle {
  idVenta: number;
  fecha: string;
  idCliente: number;
  nombreCliente: string;
  idVendedor: number;
  nombreVendedor: string;
  total: number;
  detalles: DetalleVenta[];
}

export interface CreateDetalleVentaItem {
  idProducto: number;
  cantidad: number;
}

export interface CreateVentaBody {
  idCliente: number;
  idVendedor: number;
  detalles: CreateDetalleVentaItem[];
}

export interface TotalVentasRango {
  fechaInicio: string;
  fechaFin: string;
  totalVentas: number;
}

export interface TotalPorCliente {
  idCliente: number;
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
  totalCliente: number;
}

export interface TotalPorVendedor {
  idVendedor: number;
  nombre: string;
  primerApellido: string;
  /** @nullable */
  segundoApellido?: string | null;
  totalVendedor: number;
}

export interface TotalPorProducto {
  idProducto: number;
  nombreProducto: string;
  totalPiezas: number;
  totalVendido: number;
}

export interface ResumenDashboard {
  totalVendedores: number;
  totalClientes: number;
  totalProductos: number;
  totalVentas: number;
  montoTotalMes: number;
  ventasHoy: number;
}

export type GetVentasPorRangoParams = {
  fechaInicio: string;
  fechaFin: string;
};

export type GetVentasPorClienteParams = {
  fechaInicio: string;
  fechaFin: string;
};

export type GetVentasPorVendedorParams = {
  fechaInicio: string;
  fechaFin: string;
};

export type GetVentasPorProductoParams = {
  fechaInicio: string;
  fechaFin: string;
};