# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Imparables SA - Sistema de gestión naturista
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Listar clientes con dirección
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClienteConDireccion"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Registrar cliente y dirección
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteConDireccion"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Obtener cliente por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteConDireccion"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Modificar cliente y dirección
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteConDireccion"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Eliminar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PRODUCTOS
  /productos:
    get:
      operationId: listProductos
      tags: [productos]
      summary: Listar productos
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      operationId: createProducto
      tags: [productos]
      summary: Registrar producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductoBody"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /productos/{id}:
    get:
      operationId: getProducto
      tags: [productos]
      summary: Obtener producto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProducto
      tags: [productos]
      summary: Modificar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductoBody"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProducto
      tags: [productos]
      summary: Eliminar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # VENTAS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Direccion:
      type: object
      properties:
        idDireccion:
          type: integer
        calle:
          type: string
        numero:
          type: string
        colonia:
          type: string
        ciudad:
          type: string
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
      required:
        - idDireccion
        - calle
        - numero
        - colonia
        - ciudad
    ClienteConDireccion:
      type: object
      properties:
        idCliente:
          type: integer
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
        registrado:
          type: string
        direccion:
          $ref: "#/components/schemas/Direccion"
      required:
        - idCliente
        - nombre
        - primerApellido
        - registrado
    CreateClienteBody:
      type: object
      properties:
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
        registrado:
          type: string
        calle:
          type: string
        numero:
          type: string
        colonia:
          type: string
        ciudad:
          type: string
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
      required:
        - nombre
        - primerApellido
        - registrado
        - calle
        - numero
        - colonia
        - ciudad
    Producto:
      type: object
      properties:
        idProducto:
          type: integer
        nombreProducto:
          type: string
        precio:
          type: number
      required:
        - idProducto
        - nombreProducto
        - precio
    CreateProductoBody:
      type: object
      properties:
        nombreProducto:
          type: string
        precio:
          type: number
      required:
        - nombreProducto
        - precio
...